requirements.txt
adafruit-ampy==1.1.0
alembic==1.11.1
anyio==3.6.2
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
astroid==2.15.5
astropy==5.2.2
asttokens==2.2.1
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
bilby==2.0.1
bilby.cython==0.4.1
bitstring==4.0.2
bleach==6.0.0
blinker==1.6.2
bokeh==3.1.1
Brlapi==0.8.3
cachetools==5.0.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-default-group==1.2.2
cloup==0.13.1
colorama==0.4.4
colour==0.1.5
comm==0.1.3
command-not-found==0.3
configobj==5.0.6
contourpy==1.0.7
corner==2.2.2
cryptography==40.0.2
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.34
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distro==1.7.0
Django==4.2.2
dnspython==2.3.0
docutils==0.20.1
dqsegdb2==1.1.3
dynesty==2.1.1
ecdsa==0.18.0
email-validator==2.0.0.post2
emcee==3.1.4
esptool==4.5.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
eyeD3==0.8.10
fasteners==0.14.1
fastjsonschema==2.16.3
filelock==3.12.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Reuploaded==1.3.0
Flask-Security-Too==5.2.0
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
fonttools==4.39.4
fqdn==1.5.1
gcalendar==0.4
glcontext==2.3.7
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==2.0.2
grpcio==1.30.2
gwdatafind==1.1.3
gwosc==0.7.1
gwpy==3.0.4
h5py==3.8.0
httplib2==0.20.2
idna==3.4
ifaddr==0.1.7
igwn-auth-utils==0.4.0
IMDbPY==2021.4.18
importlib-metadata==4.6.4
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
isosurfaces==0.1.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==23.5.0
kiwisolver==1.4.4
lalsuite==7.11
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
ligo-common==1.0.3
ligo-segments==1.4.0
ligotimegps==2.0.1
louis==3.20.0
lscsoft-glue==3.0.1
macaroonbakery==1.3.1
machine==0.0.1
Mako==1.2.4
manim==0.17.3
ManimPango==0.4.3
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
moderngl==5.8.2
moderngl-window==2.4.4
monotonic==1.6
more-itertools==8.10.0
mplcursors==0.5.2
mpld3==0.5.9
mpmath==1.3.0
multipledispatch==0.6.0
mypy==1.3.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nemo-emblems==5.6.0
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.8
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.0
onboard==1.4.1
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
PAM==0.4.2
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pegasus-wms.api==5.0.5
pegasus-wms.common==5.0.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.6.0
plotly==5.14.1
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pur==7.1.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.23.0
PyCBC==2.0.5
pycodestyle==2.10.0
pycparser==2.21
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.44.1
pydub==0.25.1
pyelftools==0.27
pyerfa==2.0.0.3
pyglet==2.0.7
Pygments==2.15.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pylint==2.17.4
pymacaroons==0.13.0
PyMata==2.20
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
pyrr==0.10.3
pyrsistent==0.19.3
pyRXP==3.0.1
pyserial==3.5
pytest==7.3.2
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.8
python-dotenv==1.0.0
python-gnupg==0.4.8
python-json-logger==2.0.7
python-ligo-lw==1.8.3
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
reedsolo==1.6.0
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.8
safe-netrc==1.0.1
scapy==2.5.0
scipy==1.10.1
scitokens==1.7.4
screeninfo==0.8.1
SecretStorage==3.3.1
Send2Trash==1.8.2
setproctitle==1.2.2
six==1.16.0
skia-pathops==0.7.4
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==2.0.16
sqlparse==0.4.4
srt==3.5.3
ssh-import-id==5.11
stack-data==0.6.2
svgelements==1.9.5
sympy==1.12
systemd-python==234
tenacity==8.2.2
terminado==0.17.1
thonny==4.0.2
tinycss2==1.1.1
tldextract==3.1.2
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
uri-template==1.2.0
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.23.1
wadllib==1.3.6
watchdog==2.3.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wrapt==1.15.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
xyzservices==2023.5.0
youtube-dl==2021.12.17
zipp==1.0.0
